feat: fix high+medium MVVM pending items; graduate models/widgets to features
