+++ client-app/lib/core/models/hud_notices_action_model.dart
