+++ client-app/lib/core/services/remote_config_service.dart
@@ -1,4 +1,4 @@
-import 'package:plan_sync/backend/models/remote_config/hud_notices_model.dart';
+import 'package:plan_sync/core/models/hud_notices_model.dart';
 
 abstract class RemoteConfigService {
   Future<void> onReady();
