+++ client-app/lib/features/electives/view/widgets/elective_year_bar.dart
