+++ client-app/lib/features/electives/view/widgets/electives_scheme_bar.dart
