+++ client-app/lib/features/electives/view/widgets/electives_sem_bar.dart
