+++ client-app/lib/features/home/view/widgets/report_error.dart
