+++ client-app/lib/features/schedule/model/timetable_meta.dart
