+++ client-app/lib/features/schedule/model/timetable_schedule_entry.dart
