+++ client-app/lib/features/schedule/view/widgets/sections_bar.dart
