+++ client-app/lib/features/schedule/view/widgets/semester_bar.dart
