+++ client-app/lib/features/schedule/view/widgets/year_bar.dart
