+++ client-app/lib/features/settings/view/widgets/delete_account_popup.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:plan_sync/features/auth/repository/auth_repository.dart';
+import 'package:plan_sync/features/settings/viewmodel/settings_view_model.dart';
 import 'package:plan_sync/util/snackbar.dart';
 import 'package:provider/provider.dart';
 
@@ -67,8 +68,8 @@ class _DeleteAccountPopupState extends State<DeleteAccountPopup> {
                             setState(() => isWorking = true);
                             try {
                               await context
-                                  .read<AuthRepository>()
-                                  .deleteCurrentUser();
+                                  .read<SettingsViewModel>()
+                                  .deleteAccount();
                               if (!context.mounted) return;
                               CustomSnackbar.info(
                                 'Account Deleted',
