+++ client-app/lib/features/settings/viewmodel/settings_view_model.dart
@@ -26,6 +26,8 @@ class SettingsViewModel extends ChangeNotifier {
 
   Future<void> logout() => _auth.logout();
 
+  Future<void> deleteAccount() => _auth.deleteCurrentUser();
+
   void togglePun() {
     isPunActivated = !isPunActivated;
     notifyListeners();
