+++ client-app/test/mock_controllers/remote_config_controller_mock.dart
@@ -1,5 +1,5 @@
 import 'package:mockito/mockito.dart';
-import 'package:plan_sync/backend/models/remote_config/hud_notices_model.dart';
+import 'package:plan_sync/core/models/hud_notices_model.dart';
 import 'package:plan_sync/core/services/remote_config_service.dart';
 
 class MockRemoteConfigController extends Mock implements RemoteConfigService {
