+++ client-app/test/mock_controllers/schedule_repository_mock.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:plan_sync/backend/models/timetable.dart';
+import 'package:plan_sync/features/schedule/model/timetable.dart';
 import 'package:plan_sync/features/schedule/repository/schedule_repository.dart';
 
 enum MockScheduleRepositoryStage {
