+++ client-app/test/widgets/bottom-sheets/schedule_preference_test.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
 import 'package:plan_sync/core/services/theme_service.dart';
-import 'package:plan_sync/widgets/bottom-sheets/schedule_preference.dart';
-import 'package:plan_sync/widgets/dropdowns/sections_bar.dart';
-import 'package:plan_sync/widgets/dropdowns/semester_bar.dart';
-import 'package:plan_sync/widgets/dropdowns/year_bar.dart';
+import 'package:plan_sync/features/schedule/view/widgets/schedule_preference.dart';
+import 'package:plan_sync/features/schedule/view/widgets/sections_bar.dart';
+import 'package:plan_sync/features/schedule/view/widgets/semester_bar.dart';
+import 'package:plan_sync/features/schedule/view/widgets/year_bar.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../main.dart';
