+++ client-app/test/widgets/dropdowns/elective_year_bar_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plan_sync/widgets/dropdowns/elective_year_bar.dart';
+import 'package:plan_sync/features/electives/view/widgets/elective_year_bar.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../main.dart';
