+++ client-app/test/widgets/dropdowns/electives_sem_bar_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plan_sync/widgets/dropdowns/electives_sem_bar.dart';
+import 'package:plan_sync/features/electives/view/widgets/electives_sem_bar.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../main.dart';
