+++ client-app/test/widgets/dropdowns/sections_bar_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plan_sync/widgets/dropdowns/sections_bar.dart';
+import 'package:plan_sync/features/schedule/view/widgets/sections_bar.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../main.dart';
