+++ client-app/test/widgets/time_table_for_day_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:plan_sync/backend/models/timetable.dart';
-import 'package:plan_sync/backend/models/timetable_meta.dart';
-import 'package:plan_sync/backend/models/timetable_schedule_entry.dart';
+import 'package:plan_sync/features/schedule/model/timetable.dart';
+import 'package:plan_sync/features/schedule/model/timetable_meta.dart';
+import 'package:plan_sync/features/schedule/model/timetable_schedule_entry.dart';
 import 'package:plan_sync/features/schedule/view/widgets/time_table_for_day.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../main.dart';
